Say '2 minutes ago' from one module, not two copies
