+++ app/src/lib/relative-time.ts
@@ -1,0 +1,23 @@
+const RELATIVE_UNITS = [
+  { limit: 60_000, divisor: 1_000, unit: "second" },
+  { limit: 3_600_000, divisor: 60_000, unit: "minute" },
+  { limit: 86_400_000, divisor: 3_600_000, unit: "hour" },
+  { limit: 604_800_000, divisor: 86_400_000, unit: "day" },
+  { limit: Number.POSITIVE_INFINITY, divisor: 604_800_000, unit: "week" },
+] as const;
+
+const relativeFormat = new Intl.RelativeTimeFormat(undefined, {
+  numeric: "auto",
+});
+
+/** Locale-aware relative timestamp, e.g. "2 minutes ago". */
+export function relativeTime(iso: string): string {
+  const elapsed = Date.now() - new Date(iso).getTime();
+  const scale =
+    RELATIVE_UNITS.find(({ limit }) => Math.abs(elapsed) < limit) ??
+    RELATIVE_UNITS[RELATIVE_UNITS.length - 1];
+  return relativeFormat.format(
+    -Math.round(elapsed / scale.divisor),
+    scale.unit,
+  );
+}
